Fix default implementation of AbstractSignatureArray.sizes()

+++ gambit/signatures/base.py
@@ -53,7 +53,7 @@ class AbstractSignatureArray(Sequence[KmerSignature]):
 
 	def sizes(self) -> Sequence[int]:
 		"""Get the sizes of all signatures in the array."""
-		return np.fromiter(map(self.sizeof, range(len(self))))
+		return np.fromiter(map(self.sizeof, range(len(self))), dtype=int)
 
 	@abstractmethod
 	def __getitem__(self, index: Union[int, slice, Sequence[int], Sequence[bool]]) -> Union[KmerSignature, 'AbstractSignatureArray']:
